skin-cancer-api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
tensorflow==2.18.0
Pillow==10.3.0
python-multipart==0.0.9